+++ src/main/java/fuelinjection/RentCompany.java
@@ -1,4 +1,4 @@
-package fuelInjection;
+package fuelinjection;
 
 import java.util.ArrayList;
 import java.util.List;
