+++ src/main/java/fuelinjection/Sonata.java
@@ -1,4 +1,4 @@
-package fuelInjection;
+package fuelinjection;
 
 public class Sonata extends Car {
